requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3

Werkzeug==3.0.2
stripe==15.1.0
pg8000==1.30.5
python-dotenv==1.2.2